Add files via upload
--Polar Codes NN, NNpolarCW   This is a simple fully connected net that attempts to run learn and better resolve polar codes. 

By doing your best to use the parameters provided when making modifications the CSV that is also included here gets updated. genData.py has pretty much all of the functions... it was originally intended to hold a few functions, but it kind of exploded. It's rather messy too. Sorry about that. But I tried to clean up and keep the main file commented. 

The net is fed noisy codewords and is intended to learn to produce the original codewords that can then be decoded. This seems really dumb, but it appears that due to the way polar codes function this is the most effective way.?